Combined outline and outline-node classes in same source file
